when promoting type, also promote format string
a fixup to commit 40123bec2c27 : when I promoted incount, also needed to
promote MPL_printf string to match.

fixes CID 120440

No Reviewer

+++ src/mpid/common/sched/mpid_sched.c
@@ -734,8 +734,10 @@ int MPID_Sched_copy(const void *inbuf,  MPI_Aint incount,  MPI_Datatype intype,
         MPID_Datatype_get_size_macro(intype, intype_size);
         MPID_Datatype_get_size_macro(outtype, outtype_size);
         if (incount * intype_size > outcount * outtype_size) {
-            MPL_error_printf("truncation: intype=%#x, intype_size=%lld, incount=%d, outtype=%#x, outtype_size=%lld outcount=%d\n",
-                              intype, (long long)intype_size, incount, outtype, (long long)outtype_size, outcount);
+            MPL_error_printf("truncation: intype=%#x, intype_size=%lld, incount=%lld, outtype=%#x, outtype_size=%lld outcount=%d\n",
+                              intype, (long long)intype_size,
+			      (long long)incount, outtype,
+			      (long long)outtype_size, outcount);
         }
     }
 #endif
